Add MultiplesReviewReferralResponseLegalOps
