+++ src/main/resources/wa-task-permissions-employment-et_scotland.dmn
@@ -465,7 +465,7 @@
           <text>"hearing-centre-team-leader"	</text>
         </outputEntry>
         <outputEntry id="LiteralExpression_0syfl11">
-          <text>"Read, Own, Manage, Claim, Unclaim, Assign, Unassign, Complete, Cancel"</text>
+          <text>"Assign, Unassign, Complete, Cancel"</text>
         </outputEntry>
         <outputEntry id="LiteralExpression_0rpqixd">
           <text>"ADMIN"</text>
@@ -667,7 +667,7 @@
           <text>"ctsc-team-leader"</text>
         </outputEntry>
         <outputEntry id="LiteralExpression_19yif9o">
-          <text>"Read, Own, Manage, Claim, Unclaim, Assign, Unassign, Complete, Cancel"</text>
+          <text>"Assign, Unassign, Complete, Cancel"</text>
         </outputEntry>
         <outputEntry id="LiteralExpression_1ej6vgq">
           <text>"CTSC"</text>
