+++ src/main/resources/wa-task-types-employment-et_englandwales_multiple.dmn
@@ -20,6 +20,17 @@
           <text>"Review Multiples Referral - Admin"</text>
         </outputEntry>
       </rule>
+      <rule id="DecisionRule_0d163em">
+        <inputEntry id="UnaryTests_1q9z8nx">
+          <text></text>
+        </inputEntry>
+        <outputEntry id="LiteralExpression_1uyh6qb">
+          <text>"MultiplesReviewReferralResponseLegalOps"</text>
+        </outputEntry>
+        <outputEntry id="LiteralExpression_19nvf6r">
+          <text>"Review Referral Response - Legal Ops"</text>
+        </outputEntry>
+      </rule>
     </decisionTable>
   </decision>
   <dmndi:DMNDI>
