+++ src/test/java/uk/gov/hmcts/et/taskconfiguration/dmn/EmploymentTaskConfigurationTestEW.java
@@ -25,42 +25,15 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static uk.gov.hmcts.et.taskconfiguration.DmnDecisionTable.WA_TASK_CONFIGURATION_ET_EW;
+import static uk.gov.hmcts.et.taskconfiguration.utility.ConfigurationUtility.EXTRA_TEST_CALENDAR;
+import static uk.gov.hmcts.et.taskconfiguration.utility.ConfigurationUtility.ISURGENT_REPLY_NO;
+import static uk.gov.hmcts.et.taskconfiguration.utility.ConfigurationUtility.ISURGENT_REPLY_YES;
+import static uk.gov.hmcts.et.taskconfiguration.utility.ConfigurationUtility.IS_URGENT;
+import static uk.gov.hmcts.et.taskconfiguration.utility.ConfigurationUtility.NOT_URGENT;
 
 class EmploymentTaskConfigurationTestEW extends DmnDecisionTableBaseUnitTest {
 
     private static final String DEFAULT_CALENDAR = "https://www.gov.uk/bank-holidays/england-and-wales.json";
-    private static final String EXTRA_TEST_CALENDAR = "https://raw.githubusercontent.com/hmcts/"
-        + "civil-wa-task-configuration/master/src/main/resources/privilege-calendar.json";
-
-    public static final String IS_URGENT =
-        "{\"referralCollection\": ["
-            + "{\"value\": {\"isUrgent\": \"No\",\"referralNumber\": \"1\"}},"
-            + "{\"value\": {\"isUrgent\": \"Yes\",\"referralNumber\": \"2\"}}"
-            + "]}";
-    public static final String NOT_URGENT =
-        "{\"referralCollection\": ["
-            + "{\"value\": {\"isUrgent\": \"Yes\",\"referralNumber\": \"1\"}},"
-            + "{\"value\": {\"isUrgent\": \"No\",\"referralNumber\": \"2\"}}"
-            + "]}";
-
-    public static final String ISURGENT_REPLY_YES =
-        "{\"referralCollection\":["
-            + "{\"value\":{\"referralReplyCollection\":["
-            + "{\"value\":{\"isUrgentReply\":\"No\",\"replyDateTime\":\"2023-10-01T12:00:00.00\"}},"
-            + "{\"value\":{\"isUrgentReply\":\"Yes\",\"replyDateTime\":\"2023-10-01T14:00:00.00\"}}"
-            + "]}},"
-            + "{\"value\":{\"referralReplyCollection\":["
-            + "{\"value\":{\"isUrgentReply\":\"No\",\"replyDateTime\":\"2023-10-01T13:00:00.00\"}}"
-            + "]}}]}";
-    public static final String ISURGENT_REPLY_NO =
-        "{\"referralCollection\":["
-            + "{\"value\":{\"referralReplyCollection\":["
-            + "{\"value\":{\"isUrgentReply\":\"Yes\",\"replyDateTime\":\"2023-10-01T12:00:00.00\"}},"
-            + "{\"value\":{\"isUrgentReply\":\"No\",\"replyDateTime\":\"2023-10-01T14:00:00.00\"}}"
-            + "]}},"
-            + "{\"value\":{\"referralReplyCollection\":["
-            + "{\"value\":{\"isUrgentReply\":\"Yes\",\"replyDateTime\":\"2023-10-01T13:00:00.00\"}}"
-            + "]}}]}";
 
     @BeforeAll
     public static void initialization() {
