+++ src/test/java/uk/gov/hmcts/et/taskconfiguration/dmn/multiples/EmploymentTaskCompletionMultipleTestEW.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.et.taskconfiguration.dmn;
+package uk.gov.hmcts.et.taskconfiguration.dmn.multiples;
 
 import org.camunda.bpm.dmn.engine.DmnDecisionTableResult;
 import org.camunda.bpm.dmn.engine.impl.DmnDecisionTableImpl;
@@ -18,13 +18,13 @@ import java.util.stream.Stream;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static uk.gov.hmcts.et.taskconfiguration.DmnDecisionTable.WA_TASK_COMPLETION_ET_MULTIPLE_SCOTLAND;
+import static uk.gov.hmcts.et.taskconfiguration.DmnDecisionTable.WA_TASK_COMPLETION_MULTIPLE_ET_EW;
 
-class EmploymentTaskCompletionMultipleTestScot extends DmnDecisionTableBaseUnitTest {
+class EmploymentTaskCompletionMultipleTestEW extends DmnDecisionTableBaseUnitTest {
 
     @BeforeAll
     public static void initialization() {
-        CURRENT_DMN_DECISION_TABLE = WA_TASK_COMPLETION_ET_MULTIPLE_SCOTLAND;
+        CURRENT_DMN_DECISION_TABLE = WA_TASK_COMPLETION_MULTIPLE_ET_EW;
     }
 
     static Stream<Arguments> scenarioProvider() {
@@ -35,6 +35,9 @@ class EmploymentTaskCompletionMultipleTestScot extends DmnDecisionTableBaseUnitT
                     Map.of(
                         "taskType", "ReviewReferralAdminMultiple",
                         "completionMode", "Auto"
+                    ),
+                    Map.of("taskType", "MultiplesReviewReferralResponseLegalOps",
+                           "completionMode", "Auto"
                     )
                 )
             ),
@@ -44,6 +47,9 @@ class EmploymentTaskCompletionMultipleTestScot extends DmnDecisionTableBaseUnitT
                     Map.of(
                         "taskType", "ReviewReferralAdminMultiple",
                         "completionMode", "Auto"
+                    ),
+                    Map.of("taskType", "MultiplesReviewReferralResponseLegalOps",
+                           "completionMode", "Auto"
                     )
                 )
             )
@@ -64,6 +70,6 @@ class EmploymentTaskCompletionMultipleTestScot extends DmnDecisionTableBaseUnitT
     void if_this_test_fails_needs_updating_with_your_changes() {
         //The purpose of this test is to prevent adding new rows without being tested
         DmnDecisionTableImpl logic = (DmnDecisionTableImpl) decision.getDecisionLogic();
-        assertThat(logic.getRules().size(), is(1));
+        assertThat(logic.getRules().size(), is(2));
     }
 }
